Move Harmony patches into their own directory and namespace
